+++ dwdat2py/__init__.py
@@ -17,7 +17,7 @@ import os
 import errno
 from contextlib import contextmanager
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
 CONFIGBASENAME = 'dewelibdir'